better comments and refactorings
