+++ src/internal/packager2/inspect.go
@@ -55,10 +55,8 @@ func InspectPackageResources(ctx context.Context, pkgLayout *layout2.PackageLayo
 	if err != nil {
 		return InspectPackageResourcesResults{}, err
 	}
-	// TODO make this a callable function
-	variableConfig := template.GetZarfVariableConfig(ctx)
-	variableConfig.SetConstants(pkgLayout.Pkg.Constants)
-	if err := variableConfig.PopulateVariables(pkgLayout.Pkg.Variables, opts.SetVariables); err != nil {
+	variableConfig, err := getPopulatedVariableConfig(ctx, pkgLayout.Pkg, opts.SetVariables)
+	if err != nil {
 		return InspectPackageResourcesResults{}, err
 	}
 	tmpPackagePath, err := utils.MakeTempDir(config.CommonOptions.TempDirectory)
@@ -95,14 +93,9 @@ func InspectPackageResources(ctx context.Context, pkgLayout *layout2.PackageLayo
 			}
 
 			for _, chart := range component.Charts {
-				chartOverrides := make(map[string]any)
-				for _, variable := range chart.Variables {
-					if setVar, ok := variableConfig.GetSetVariable(variable.Name); ok && setVar != nil {
-						// Use the variable's path as a key to ensure unique entries for variables with the same name but different paths.
-						if err := helpers.MergePathAndValueIntoMap(chartOverrides, variable.Path, setVar.Value); err != nil {
-							return InspectPackageResourcesResults{}, fmt.Errorf("unable to merge path and value into map: %w", err)
-						}
-					}
+				chartOverrides, err := generateValuesOverrides(chart, component.Name, variableConfig, nil)
+				if err != nil {
+					return InspectPackageResourcesResults{}, err
 				}
 				if err := templateValuesFiles(chart, valuesDir, variableConfig); err != nil {
 					return InspectPackageResourcesResults{}, err
