+++ src/test/e2e/34_custom_init_package_test.go
@@ -38,7 +38,7 @@ func TestCustomInit(t *testing.T) {
 	// Test that we get an error when trying to deploy a package without providing the public key
 	stdOut, stdErr, err = e2e.Zarf(t, "init", "--confirm")
 	require.Error(t, err, stdOut, stdErr)
-	require.Contains(t, stdErr, "unable to load package: package is signed but no key was provided - add a key with the --key flag or use the --skip-signature-validation flag and run the command again")
+	require.Contains(t, stdErr, "unable to load package: package is signed but no key was provided")
 
 	/* Test operations during package deploy */
 	// Test that we can deploy the package with the public key
